♻️ refactor: 페이지 주소 변경
